# performance: minimize update calls