--- a/pkg/objectstate/api/api.go
+++ b/pkg/objectstate/api/api.go
@@ -29,6 +29,7 @@ import (
 	"github.com/openshift-kni/numaresources-operator/pkg/objectstate"
 	"github.com/openshift-kni/numaresources-operator/pkg/objectstate/compare"
 	"github.com/openshift-kni/numaresources-operator/pkg/objectstate/merge"
+	"github.com/openshift-kni/numaresources-operator/pkg/objectstate/setdefault"
 )
 
 type ExistingManifests struct {
@@ -43,7 +44,8 @@ func (em *ExistingManifests) State(mf apimanifests.Manifests) []objectstate.Obje
 			Error:    em.CrdError,
 			Desired:  mf.Crd.DeepCopy(),
 			Compare:  compare.Object,
-			Merge:    merge.MetadataForUpdate,
+			Merge:    merge.ObjectForUpdate,
+			Default:  setdefault.CRD,
 		},
 	}
 }