--- a/pkg/objectstate/cfg/cfg.go
+++ b/pkg/objectstate/cfg/cfg.go
@@ -26,6 +26,7 @@ import (
 	"github.com/openshift-kni/numaresources-operator/pkg/objectstate"
 	"github.com/openshift-kni/numaresources-operator/pkg/objectstate/compare"
 	"github.com/openshift-kni/numaresources-operator/pkg/objectstate/merge"
+	"github.com/openshift-kni/numaresources-operator/pkg/objectstate/setdefault"
 )
 
 type Manifests struct {
@@ -45,6 +46,7 @@ func (em *ExistingManifests) State(mf Manifests) []objectstate.ObjectState {
 			Desired:  mf.Config.DeepCopy(),
 			Compare:  compare.Object,
 			Merge:    merge.ObjectForUpdate,
+			Default:  setdefault.None,
 		},
 	}
 }