--- a/pkg/objectstate/rte/machineconfigpool.go
+++ b/pkg/objectstate/rte/machineconfigpool.go
@@ -32,6 +32,7 @@ import (
 	"github.com/openshift-kni/numaresources-operator/pkg/objectstate"
 	"github.com/openshift-kni/numaresources-operator/pkg/objectstate/compare"
 	"github.com/openshift-kni/numaresources-operator/pkg/objectstate/merge"
+	"github.com/openshift-kni/numaresources-operator/pkg/objectstate/setdefault"
 )
 
 type machineConfigPoolFinder struct {
@@ -116,6 +117,7 @@ func (obj machineConfigPoolFinder) FindState(mf Manifests, tree nodegroupv1.Tree
 			Desired:     desiredDaemonSet,
 			Compare:     compare.Object,
 			Merge:       merge.ObjectForUpdate,
+			Default:     setdefault.DaemonSet,
 		})
 	}
 	return ret