--- a/pkg/objectstate/rte/nodegroup.go
+++ b/pkg/objectstate/rte/nodegroup.go
@@ -30,6 +30,7 @@ import (
 	"github.com/openshift-kni/numaresources-operator/pkg/objectstate"
 	"github.com/openshift-kni/numaresources-operator/pkg/objectstate/compare"
 	"github.com/openshift-kni/numaresources-operator/pkg/objectstate/merge"
+	"github.com/openshift-kni/numaresources-operator/pkg/objectstate/setdefault"
 )
 
 type nodeGroupFinder struct {
@@ -100,6 +101,7 @@ func (obj nodeGroupFinder) FindState(mf Manifests, tree nodegroupv1.Tree) []obje
 		Desired:     desiredDaemonSet,
 		Compare:     compare.Object,
 		Merge:       merge.ObjectForUpdate,
+		Default:     setdefault.DaemonSet,
 	})
 	return ret
 }