--- a/pkg/objectstate/rte/rte.go
+++ b/pkg/objectstate/rte/rte.go
@@ -38,6 +38,7 @@ import (
 	"github.com/openshift-kni/numaresources-operator/pkg/objectstate"
 	"github.com/openshift-kni/numaresources-operator/pkg/objectstate/compare"
 	"github.com/openshift-kni/numaresources-operator/pkg/objectstate/merge"
+	"github.com/openshift-kni/numaresources-operator/pkg/objectstate/setdefault"
 )
 
 const (
@@ -155,6 +156,7 @@ func (em *ExistingManifests) MachineConfigsState(mf Manifests) ([]objectstate.Ob
 					Desired:  desiredMachineConfig,
 					Compare:  compare.Object,
 					Merge:    merge.ObjectForUpdate,
+					Default:  setdefault.None,
 				},
 			)
 			enabledMCCount++
@@ -244,6 +246,7 @@ func (em *ExistingManifests) State(mf Manifests) []objectstate.ObjectState {
 			Desired:  mf.Core.ServiceAccount.DeepCopy(),
 			Compare:  compare.Object,
 			Merge:    merge.ServiceAccountForUpdate,
+			Default:  setdefault.None,
 		},
 		{
 			Existing: em.existing.Core.Role,
@@ -251,6 +254,7 @@ func (em *ExistingManifests) State(mf Manifests) []objectstate.ObjectState {
 			Desired:  mf.Core.Role.DeepCopy(),
 			Compare:  compare.Object,
 			Merge:    merge.ObjectForUpdate,
+			Default:  setdefault.None,
 		},
 		{
 			Existing: em.existing.Core.RoleBinding,
@@ -258,6 +262,7 @@ func (em *ExistingManifests) State(mf Manifests) []objectstate.ObjectState {
 			Desired:  mf.Core.RoleBinding.DeepCopy(),
 			Compare:  compare.Object,
 			Merge:    merge.ObjectForUpdate,
+			Default:  setdefault.RoleBinding,
 		},
 		{
 			Existing: em.existing.Core.ClusterRole,
@@ -265,6 +270,7 @@ func (em *ExistingManifests) State(mf Manifests) []objectstate.ObjectState {
 			Desired:  mf.Core.ClusterRole.DeepCopy(),
 			Compare:  compare.Object,
 			Merge:    merge.ObjectForUpdate,
+			Default:  setdefault.None,
 		},
 		{
 			Existing: em.existing.Core.ClusterRoleBinding,
@@ -272,6 +278,7 @@ func (em *ExistingManifests) State(mf Manifests) []objectstate.ObjectState {
 			Desired:  mf.Core.ClusterRoleBinding.DeepCopy(),
 			Compare:  compare.Object,
 			Merge:    merge.ObjectForUpdate,
+			Default:  setdefault.ClusterRoleBinding,
 		},
 	}
 
@@ -282,6 +289,7 @@ func (em *ExistingManifests) State(mf Manifests) []objectstate.ObjectState {
 			Desired:  mf.Core.SecurityContextConstraint.DeepCopy(),
 			Compare:  compare.Object,
 			Merge:    merge.ObjectForUpdate,
+			Default:  setdefault.SecurityContextConstraints,
 		})
 	}
 	if mf.Core.SecurityContextConstraintV2 != nil {
@@ -291,6 +299,7 @@ func (em *ExistingManifests) State(mf Manifests) []objectstate.ObjectState {
 			Desired:  mf.Core.SecurityContextConstraintV2.DeepCopy(),
 			Compare:  compare.Object,
 			Merge:    merge.ObjectForUpdate,
+			Default:  setdefault.SecurityContextConstraints,
 		})
 	}
 
@@ -308,6 +317,7 @@ func (em *ExistingManifests) State(mf Manifests) []objectstate.ObjectState {
 		Desired:  mf.Metrics.Service.DeepCopy(),
 		Compare:  compare.Object,
 		Merge:    merge.MetadataForUpdate,
+		Default:  setdefault.Service,
 	})
 
 	return ret