Many of the `Update` calls done by the operator
are changing nothing.

This is happening because the API server defaults some of the fields
while our code leave them empty.

This result in a diff in the comparison function and leads to redundant calls.

This package purpose it to set the default values in our objects similar to what
the API server is doing.

The code is based on upstream code that is
not in staging so we can't import it unfortunately.

Below benchmarks shows the benefit of those changes:

**Simple patch against `numaresources` CR (before change):**
```
oc patch numaresourcesoperators.nodetopology.openshift.io numaresourcesoperator \
  --type='json' -p '[{"op": "add", "path": "/spec/nodeGroups/0/annotations", "value": {"key3":"value6"}}]'
numaresourcesoperator.nodetopology.openshift.io/numaresourcesoperator patched
```
```
./kubectl-dev_tool audit -f audit.log --after 2025-02-18T17:10:00+02:00 -otop --by resource --verb=update --verb=create  --user=system:serviceaccount:openshift-numaresources:numaresources-controller-manager  --resource=*.* --resource=-leases.* 
count: 7, first: 2025-02-18T17:12:26+02:00, last: 2025-02-18T17:12:26+02:00, duration: 59.093ms
2x                   security.openshift.io/v1/securitycontextconstraints
1x                   apiextensions.k8s.io/v1/customresourcedefinitions
1x                   nodetopology.openshift.io/v1/numaresourcesoperators
1x                   v1/services
1x                   apps/daemonsets
1x                   v1/serviceaccounts
```

**Simple patch against `numaresources` CR (after change):**
```
oc patch numaresourcesoperators.nodetopology.openshift.io numaresourcesoperator \
  --type='json' -p '[{"op": "add", "path": "/spec/nodeGroups/0/annotations", "value": {"key3":"value6"}}]'
numaresourcesoperator.nodetopology.openshift.io/numaresourcesoperator patched
```
```
./kubectl-dev_tool audit -f audit.log --after 2025-02-18T15:30:00+02:00 -otop --by resource --verb=update --verb=create  --user=system:serviceaccount:openshift-numaresources:numaresources-controller-manager  --resource=*.* --resource=-leases.* 
count: 1, first: 2025-02-18T17:02:34+02:00, last: 2025-02-18T17:02:34+02:00, duration: 0s
1x                   nodetopology.openshift.io/v1/numaresourcesoperators
```

**Realistic benchmark against serial suite (before change):**
```
E2E_NROP_INSTALL_SKIP_KC=true ./bin/e2e-nrop-serial.test  --ginkgo.skip='reboot_required|schedrst' --ginkgo.label-filter='!(feature: containsAny { ngpoolname, nganns })' --ginkgo.label-filter='tier0||tier1' --ginkgo.dry-run

Ran 32 of 120 Specs in 2895.370 seconds
SUCCESS! -- 32 Passed | 0 Failed | 0 Pending | 88 Skipped
```
```
./kubectl-dev_tool audit -f audit.log --after 2025-02-18T14:10:00+02:00 -otop --by resource --verb=update --verb=create --user=system:serviceaccount:openshift-numaresources:numaresources-controller-manager --resource=*.* --resource=-leases.* 
count: 113, first: 2025-02-18T14:40:20+02:00, last: 2025-02-18T15:07:09+02:00, duration: 26m48.655819s
32x                  security.openshift.io/v1/securitycontextconstraints
16x                  apps/daemonsets
16x                  v1/serviceaccounts
16x                  apiextensions.k8s.io/v1/customresourcedefinitions
16x                  nodetopology.openshift.io/v1/numaresourcesoperators
16x                  v1/services
1x                   v1/events
```

**Realistic benchmark against serial suite (after change):**

```
E2E_NROP_INSTALL_SKIP_KC=true ./bin/e2e-nrop-serial.test  --ginkgo.skip='reboot_required|schedrst' --ginkgo.label-filter='!(feature: containsAny { ngpoolname, nganns })' --ginkgo.label-filter='tier0||tier1' --ginkgo.dry-run

Ran 32 of 120 Specs in 2901.802 seconds
SUCCESS! -- 32 Passed | 0 Failed | 0 Pending | 88 Skipped
```

```
./kubectl-dev_tool audit -f audit.log --after 2025-02-18T16:00:00+02:00 -otop --by resource --verb=update --verb=create --user=system:serviceaccount:openshift-numaresources:numaresources-controller-manager --resource=*.* --resource=-leases.* 

// return empty. no redundant call at all
```

**NOTE:** we must add a CI lane that tracks the amount of update call and failed if we're exceeding a certain limit.
This is needed because atm we don't have a better way to track changes in the set-default  logic of the API server.