@ffromani 2025-02-19
some historical context.
The person who wrote the merge code (cough cough) implemented the bare minimum merge code back in time. That was not very smart and admittedly pretty raw. The thought was like "we will optimize later" and later, of course, never came.

I'm not against using `defaulter`. This is actually a good solution and another very valid approach which complements smart and precise merge functions. The problem is having easy access to that package.