@ffromani 2025-05-12
reviewing this PR is high on my priority. This is very good work. Conceptually I want this in, let's see how we can address the implementation details.