@ffromani 2025-05-26
I spent most of the morning reviewing this PR and trying to remove the `k8s.io/kubernetes` dependency. It's quite hard.
At the same time, reintroducing the dep on `k8s.io/kubernetes` is something we can't do. Despite how good this is PR, the ROI is still not high enough.
What we can do to move forward is to extract the few changes which don't depend on the `k8s.io/kubernetes` (directly or indirectly) and start merging those. I think 1371710b (plus: ```
diff --git a/pkg/objectstate/merge/merge.go b/pkg/objectstate/merge/merge.go
index 27c8119b..8ccf4df6 100644
--- a/pkg/objectstate/merge/merge.go
+++ b/pkg/objectstate/merge/merge.go
@@ -21,7 +21,7 @@ import (
 
        appsv1 "k8s.io/api/apps/v1"
        corev1 "k8s.io/api/core/v1"
-       v1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
+       apiextv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
        "k8s.io/apimachinery/pkg/util/sets"
 
        "sigs.k8s.io/controller-runtime/pkg/client"
@@ -62,8 +62,8 @@ func StatusForUpdate(current client.Object, updated client.Object) (client.Objec
                updated.(*appsv1.DaemonSet).Status = currentTyped.Status
        case *corev1.Service:
                updated.(*corev1.Service).Status = currentTyped.Status
-       case *v1.CustomResourceDefinition:
-               updated.(*v1.CustomResourceDefinition).Status = currentTyped.Status
+       case *apiextv1.CustomResourceDefinition:
+               updated.(*apiextv1.CustomResourceDefinition).Status = currentTyped.Status
        default:
                return updated, nil
        }
```
and 1c64a334 are easy to go in. Everything else needs more thought