@ffromani 2025-05-26
https://github.com/openshift-kni/numaresources-operator/pull/1194/commits/528cad64c4822601654c38815edbf276ce7bfac3 and https://github.com/openshift-kni/numaresources-operator/pull/1194/commits/57425ccd5e31ffb8fbb20ead0a27091de895a1f3 we can merge quickly