@ffromani 2025-06-06
I spent few hours looking at the generator script.
The problem is the script is out of necessity complex and does many transformations to the the original source code to create our packages. I see why it is like this. I can't think of a better way. Still, this is too complex and fragile to be maintained. We can extract some more bits from this PR (https://github.com/openshift-kni/numaresources-operator/pull/1313) but I think the generator script per se cannot move forward, so the `defaulter` idea can't progress.

However.
Thinking hard and (relatively) long about how to achieve the same result without the defaulter, there is a relatively contained possible way reorganizing how we do the merge step.

The merge step is how we compute the delta between the desired state and the current state, which is what we use as payload of our `Update` calls.
The existing state is what we got from the cluster (a glorified set of `Get` calls hidden in `FromClient` calls).
The desired state is created by deployer libraries:
1. starting from zeroed objects
2. enforcing the values we want, optionally mutated by the config

Thinking about it, we already have our delta, but implicit and buried in the deployer libs.
If the deployer libs:
1. make all the changes they need to manifests programmatically, instead of relying on loaded data (implementation detail)
2. can take a random object as input to mutate, vs implicitly and always use the loaded data as starting point

then we can
1. use the existing data as base
2. apply mutations from deployer libs to reach the desired state
3. use this data (plus optimization we have already maybe) as payload of `Update` calls.

this makes our `merge` package useless, replaced by functionality from deployer libs, but should be doable with targeted and contained changes in our controller logic.