@Tal-or 2025-07-01
/close superseded by #1502 which implements  https://github.com/openshift-kni/numaresources-operator/pull/1194#issuecomment-2948265049